requirements.txt
# Exchange connectivity
ccxt>=4.3.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Technical analysis
ta>=0.11.0

# Vectorised backtesting
vectorbt>=0.26.1

# AI / Claude
anthropic>=0.28.0

# Config & environment
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Market data
yfinance>=0.2.40

# Async networking
aiohttp>=3.9.0
websockets>=12.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Telegram bot (for telegram_bridge.py)
python-telegram-bot>=21.0

# Visualization
plotly>=5.22.0

# Dev / testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0

# Research module
feedparser>=6.0.0         # RSS/Atom feed parsing (news_ingest)
requests>=2.32.0          # HTTP client for news/macro APIs
pandas_ta>=0.3.14b        # Optional: more accurate TA indicators (fundamentals.py)
